+++ src/main/java/roomescape/controller/view/MemberController.java
@@ -1,10 +1,10 @@
-package roomescape.controller;
+package roomescape.controller.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class MemberPageController {
+public class MemberController {
     @GetMapping("/reservation")
     public String findReservationPage() {
         return "reservation";
