+++ src/test/java/roomescape/controller/reservation/RoomThemeControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.reservation;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
